FarmAI/requirements.txt
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
flask-msearch==0.2.6
flask-ngrok==0.0.25
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gunicorn==20.0.4
itsdangerous==1.1.0
json5==0.8.5
Keras==2.2.4
matplotlib==3.1.1
numpy==1.17.2
opencv-python==4.1.2.30
pandas==0.25.1
Pillow==6.2.0
requests==2.22.0
scikit-learn==0.21.3
SQLAlchemy==1.3.9
statsmodels==0.10.1
tensorflow==1.14.0
urllib3==1.24.2
virtualenv==20.0.20
WTForms==2.2.1
wurlitzer==1.0.3
Werkzeug==0.16.1